+++ sdk/communication/communication-identity/recordings/node/exchange_teams_token_for_acs_token_playbacklive/recording_successfully_exchanges_a_teams_token_for_an_acs_token.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "b305e4aa792cd5394b4926d36eff9389";
+module.exports.hash = "040e4a6b70cd13b74637e54d2b718c73";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}