+++ sdk/communication/communication-identity/recordings/node/exchange_teams_token_for_acs_token_playbacklive_aad/recording_throws_an_error_when_attempting_to_exchange_an_invalid_teams_token.js
@@ -1,5 +1,5 @@
 let nock = require('nock');
 
-module.exports.hash = "f91d4ed33615271226be0796624a6e88";
+module.exports.hash = "b9667d5a0f669eb1b53db67209046ed6";
 
 module.exports.testInfo = {"uniqueName":{},"newDate":{}}